requirements.txt
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
jwt==1.2.0
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pycparser==2.20
redis==3.5.3
Werkzeug==1.0.1
